+++ src/main/java/com/admir/demiraj/datacatalogspringboot/repository/FunctionsRepository.java
@@ -9,11 +9,13 @@ import com.admir.demiraj.datacatalogspringboot.resources.Functions;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
+import java.math.BigInteger;
+
 /**
  *
  * @author root
  */
 @CrossOrigin(origins = "http://localhost:4200")
-public interface FunctionsRepository extends JpaRepository<Functions, Long>{
+public interface FunctionsRepository extends JpaRepository<Functions, BigInteger>{
     
 }
